Moved Server struct to server::ftpserver (#166)
